requirements.txt
google-generativeai>=0.3.2
pillow>=10.0.0
pyautogui>=0.9.54
ttkbootstrap>=1.10.1
python-dotenv>=1.0.0
numpy>=1.24.0
packaging>=23.0
requests>=2.31.0 